+++ apps/frontend/src/lib/components/blocks/field-control/date-control.svelte
@@ -14,6 +14,7 @@
   export let readonly = false
   export let disabled = false
   export let field: DateField
+  export let onValueChange: (value: string | Date | undefined | null) => void
 
   $: formatter = field.formatter
   $: includeTime = field.includeTime
@@ -32,10 +33,11 @@
 
   let open = false
 
-  function onChange() {
+  function onChange(value: string | Date | undefined | null) {
     if (!includeTime) {
       open = false
     }
+    onValueChange?.(value)
   }
 </script>
 
@@ -61,8 +63,8 @@
   <Popover.Content class="p-0" side="bottom" align="start">
     <div class="p-1">
       <DateMacroPicker
-        onValueChange={() => {
-          onChange()
+        onValueChange={(v) => {
+          onChange(v)
         }}
         bind:value
       />
@@ -70,12 +72,14 @@
     <Calendar
       value={isString(internalDate) && isDateFieldMacro(internalDate) ? undefined : internalDate}
       onValueChange={(v) => {
+        let vv
         if (v) {
-          value = v.toString()
+          vv = v.toString()
         } else {
-          value = undefined
+          vv = undefined
         }
-        onChange()
+        value = vv
+        onChange(vv)
       }}
       initialFocus
     />
@@ -88,7 +92,9 @@
           }}
           onValueChange={(v) => {
             if (!value) return
-            value = new Date(new Date(value).setHours(v.hour, v.minute, 0, 0)).toISOString()
+            const vv = new Date(new Date(value).setHours(v.hour, v.minute, 0, 0)).toISOString()
+            value = vv
+            onValueChange?.(vv)
           }}
         />
       </div>
@@ -98,8 +104,9 @@
         class="flex-1"
         variant="outline"
         on:click={() => {
-          value = today(getLocalTimeZone()).toString()
-          onChange()
+          const v = today(getLocalTimeZone()).toString()
+          value = v
+          onChange(v)
         }}>Today</Button
       >
       <Button
@@ -108,6 +115,7 @@
         on:click={() => {
           if (value) {
             value = null
+            onValueChange?.(null)
           }
           open = false
         }}