+++ apps/frontend/src/lib/components/blocks/field-control/number-control.svelte
@@ -3,8 +3,11 @@
 
   export let readonly = false
   export let value: number
+  export let onValueChange: (value: number) => void
   const onInput = (event: Event) => {
-    value = +(event.target as HTMLInputElement).value
+    const v = Number((event.target as HTMLInputElement).value)
+    value = v
+    onValueChange?.(v)
   }
 </script>