+++ apps/frontend/src/lib/components/blocks/field-options/button-field-option.svelte
@@ -17,9 +17,9 @@
   import { Button } from "$lib/components/ui/button"
   import Separator from "$lib/components/ui/separator/separator.svelte"
   import { writable } from "svelte/store"
-  import type { ZodUndefined } from "@undb/zod"
+  import { Switch } from "$lib/components/ui/switch"
   import FiltersEditor from "../filters-editor/filters-editor.svelte"
-  import { onMount } from "svelte"
+  import { onMount, tick } from "svelte"
 
   const table = getTable()
 
@@ -40,15 +40,18 @@
       confirm: true,
     },
   }
-  const value = writable<MaybeConditionGroup<ZodUndefined> | undefined>()
+  const value = writable<MaybeConditionGroup<any> | undefined>()
   $: validValue = $table && $value ? parseValidViewFilter($table.schema, $value) : undefined
   $: if (validValue) {
     option.disabled = validValue
   }
 
+  let disabledWhen = false
+
   onMount(() => {
     if (option.disabled) {
       value.set(toMaybeConditionGroup(option.disabled))
+      disabledWhen = true
     }
   })
 
@@ -63,14 +66,19 @@
   <Label for="label">Label</Label>
   <Input class="w-full" placeholder="Button" id="label" bind:value={option.label} />
 
-  <div class="space-y-2 rounded-sm border pt-2">
-    <Label class="pl-4 text-xs font-semibold" for="disabled">Disabled When...</Label>
-    <FiltersEditor
-      bind:value={$value}
-      table={$table}
-      filter={(field) => visibleFields.some((f) => f.id.value === field.id) && getIsFilterableFieldType(field.type)}
-    ></FiltersEditor>
-  </div>
+  <Label class="flex items-center gap-2 text-xs font-semibold" for="disabled">
+    <Switch id="disabled" bind:checked={disabledWhen} size="sm" />
+    Disabled When...
+  </Label>
+  {#if disabledWhen}
+    <div class="space-y-2 rounded-sm border pt-2">
+      <FiltersEditor
+        bind:value={$value}
+        table={$table}
+        filter={(field) => visibleFields.some((f) => f.id.value === field.id) && getIsFilterableFieldType(field.type)}
+      ></FiltersEditor>
+    </div>
+  {/if}
 
   <div class="space-y-2 rounded-md border px-4 py-3">
     <p class="text-xs font-semibold">Update Value when Click Button</p>
@@ -79,7 +87,13 @@
         value.field && $table ? $table.schema.getFieldById(new FieldIdVo(value.field)).unwrap() : undefined}
       <FieldPicker
         class="w-full"
-        bind:value={value.field}
+        value={value.field}
+        onValueChange={async (field) => {
+          value.field = field
+          value.value = undefined
+          option = option
+          await tick()
+        }}
         {disabled}
         filter={(f) =>
           getIsMutableFieldType(f.type) &&
@@ -90,7 +104,12 @@
         <FieldControl
           class="text-xs"
           placeholder="Value to update..."
-          bind:value={value.value}
+          value={value.value}
+          onValueChange={async (v) => {
+            value.value = v
+            option = option
+            await tick()
+          }}
           {field}
           tableId={$table?.id.value}
         />