+++ packages/table/src/table.do.ts
@@ -128,7 +128,8 @@ export class TableDo extends AggregateRoot<ITableEvents> {
     const visibleFields = this.getViewFields(viewId).getVisibleFields()
 
     return visibleFields
-      .map((fieldId) => this.schema.getFieldById(new FieldIdVo(fieldId)).unwrap())
+      .map((fieldId) => this.schema.getFieldById(new FieldIdVo(fieldId)).into(undefined))
+      .filter((f) => !!f)
       .filter((field) => (view.showSystemFields ? true : !field.isSystem))
   }
 
@@ -182,6 +183,12 @@ export class TableDo extends AggregateRoot<ITableEvents> {
     if (view?.type === "kanban" || view?.type === "calendar" || view?.type === "gallery") {
       const fieldId = view.field.into(undefined)
       if (fieldId) {
+        addField(fieldId)
+      }
+    }
+    if (view?.color.isSome()) {
+      const fieldIds = view.color.unwrap().fieldIds
+      for (const fieldId of fieldIds) {
         addField(fieldId)
       }
     }